@@ -101,9 +82,9 @@
     <yarn-berry.version>4.0.2</yarn-berry.version>
     <yarn-berry.sha256sum>825003a0f561ad09a3b1ac4a3b3ea6207af2796d54f62a9420520915721f5186</yarn-berry.sha256sum>
 
-    <jjwt.version>0.12.5</jjwt.version>
-    <testcontainers.version>1.19.7</testcontainers.version>
-    <spring-boot.version>3.2.4</spring-boot.version>
+    <jjwt.version>0.12.6</jjwt.version>
+    <testcontainers.version>1.20.2</testcontainers.version>
+    <spring-boot.version>3.3.4</spring-boot.version>
   </properties>
 
   <dependencyManagement>
