@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c) 2023 Jenkins Infra
+ * Copyright (c) 2023-2024 Jenkins Infra
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -21,7 +21,6 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-
 package io.jenkins.pluginhealth.scoring.probes;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -60,47 +59,45 @@ class KnownSecurityVulnerabilityProbeTest extends AbstractProbeTest<KnownSecurit
         final ProbeContext ctx = mock(ProbeContext.class);
         final KnownSecurityVulnerabilityProbe probe = getSpy();
 
-        when(ctx.getUpdateCenter()).thenReturn(
-            new UpdateCenter(
-                Collections.emptyMap(),
-                Collections.emptyMap(),
-                Collections.emptyList()
-            )
-        );
+        when(ctx.getUpdateCenter())
+                .thenReturn(new UpdateCenter(Collections.emptyMap(), Collections.emptyMap(), Collections.emptyList()));
 
         final ProbeResult result = probe.apply(plugin, ctx);
         assertThat(result)
-            .isNotNull()
-            .usingRecursiveComparison()
-            .comparingOnlyFields("id", "status", "message")
-            .isEqualTo(ProbeResult.success(KnownSecurityVulnerabilityProbe.KEY, "No known security vulnerabilities.", probe.getVersion()));
+                .isNotNull()
+                .usingRecursiveComparison()
+                .comparingOnlyFields("id", "status", "message")
+                .isEqualTo(ProbeResult.success(
+                        KnownSecurityVulnerabilityProbe.KEY, "No known security vulnerabilities.", probe.getVersion()));
     }
 
     @Test
     void shouldBeOKWithWarningOnDifferentPlugin() {
         final String pluginName = "foo-bar";
         final VersionNumber pluginVersion = new VersionNumber("1.0");
-        final var pluginInUC = new Plugin(pluginName, pluginVersion, "scm", ZonedDateTime.now().minusHours(1), List.of(), 0, "", "main");
+        final var pluginInUC = new Plugin(
+                pluginName, pluginVersion, "scm", ZonedDateTime.now().minusHours(1), List.of(), 0, "", "main");
         final var plugin = mock(io.jenkins.pluginhealth.scoring.model.Plugin.class);
         final ProbeContext ctx = mock(ProbeContext.class);
         final KnownSecurityVulnerabilityProbe probe = getSpy();
 
-        when(ctx.getUpdateCenter()).thenReturn(
-            new UpdateCenter(
-                Map.of(pluginName, pluginInUC),
-                Collections.emptyMap(),
-                List.of(
-                    new SecurityWarning("SECURITY-1", "wiz", List.of(new SecurityWarningVersion(null, ".*")))
-                )
-            )
-        );
+        when(ctx.getUpdateCenter())
+                .thenReturn(new UpdateCenter(
+                        Map.of(pluginName, pluginInUC),
+                        Collections.emptyMap(),
+                        List.of(new SecurityWarning(
+                                "SECURITY-1",
+                                "wiz",
+                                "http://link-to-issue",
+                                List.of(new SecurityWarningVersion(null, ".*"))))));
 
         final ProbeResult result = probe.apply(plugin, ctx);
         assertThat(result)
-            .isNotNull()
-            .usingRecursiveComparison()
-            .comparingOnlyFields("id", "status", "message")
-            .isEqualTo(ProbeResult.success(KnownSecurityVulnerabilityProbe.KEY, "No known security vulnerabilities.", probe.getVersion()));
+                .isNotNull()
+                .usingRecursiveComparison()
+                .comparingOnlyFields("id", "status", "message")
+                .isEqualTo(ProbeResult.success(
+                        KnownSecurityVulnerabilityProbe.KEY, "No known security vulnerabilities.", probe.getVersion()));
     }
 
     @Test
@@ -114,24 +111,23 @@ class KnownSecurityVulnerabilityProbeTest extends AbstractProbeTest<KnownSecurit
 
         when(plugin.getName()).thenReturn(pluginName);
         when(plugin.getVersion()).thenReturn(pluginVersion);
-        when(ctx.getUpdateCenter()).thenReturn(
-            new UpdateCenter(
-                Collections.emptyMap(),
-                Collections.emptyMap(),
-                List.of(
-                    new SecurityWarning("SECURITY-1", pluginName, List.of(
-                        new SecurityWarningVersion(new VersionNumber("1.0"), "0\\.*")
-                    ))
-                )
-            )
-        );
+        when(ctx.getUpdateCenter())
+                .thenReturn(new UpdateCenter(
+                        Collections.emptyMap(),
+                        Collections.emptyMap(),
+                        List.of(new SecurityWarning(
+                                "SECURITY-1",
+                                pluginName,
+                                "http://link-to-issue",
+                                List.of(new SecurityWarningVersion(new VersionNumber("1.0"), "0\\.*"))))));
 
         final ProbeResult result = probe.apply(plugin, ctx);
         assertThat(result)
-            .isNotNull()
-            .usingRecursiveComparison()
-            .comparingOnlyFields("id", "status", "message")
-            .isEqualTo(ProbeResult.success(KnownSecurityVulnerabilityProbe.KEY, "No known security vulnerabilities.", probe.getVersion()));
+                .isNotNull()
+                .usingRecursiveComparison()
+                .comparingOnlyFields("id", "status", "message")
+                .isEqualTo(ProbeResult.success(
+                        KnownSecurityVulnerabilityProbe.KEY, "No known security vulnerabilities.", probe.getVersion()));
     }
 
     @Test
@@ -139,6 +135,7 @@ class KnownSecurityVulnerabilityProbeTest extends AbstractProbeTest<KnownSecurit
         final String pluginName = "foo-bar";
         final VersionNumber pluginVersion = new VersionNumber("1.1");
         final String warningId = "SECURITY-1";
+        final String url = "http://link-to-issue";
 
         final var plugin = mock(io.jenkins.pluginhealth.scoring.model.Plugin.class);
         final ProbeContext ctx = mock(ProbeContext.class);
@@ -146,22 +143,23 @@ class KnownSecurityVulnerabilityProbeTest extends AbstractProbeTest<KnownSecurit
 
         when(plugin.getName()).thenReturn(pluginName);
         when(plugin.getVersion()).thenReturn(pluginVersion);
-        when(ctx.getUpdateCenter()).thenReturn(
-            new UpdateCenter(
-                Collections.emptyMap(),
-                Collections.emptyMap(),
-                List.of(
-                    new SecurityWarning(warningId, pluginName, List.of(new SecurityWarningVersion(pluginVersion, "1.0")))
-                )
-            )
-        );
+        when(ctx.getUpdateCenter())
+                .thenReturn(new UpdateCenter(
+                        Collections.emptyMap(),
+                        Collections.emptyMap(),
+                        List.of(new SecurityWarning(
+                                warningId,
+                                pluginName,
+                                url,
+                                List.of(new SecurityWarningVersion(pluginVersion, "1.0"))))));
 
         final ProbeResult result = probe.apply(plugin, ctx);
         assertThat(result)
-            .isNotNull()
-            .usingRecursiveComparison()
-            .comparingOnlyFields("id", "status", "message")
-            .isEqualTo(ProbeResult.success(KnownSecurityVulnerabilityProbe.KEY, warningId, probe.getVersion()));
+                .isNotNull()
+                .usingRecursiveComparison()
+                .comparingOnlyFields("id", "status", "message")
+                .isEqualTo(ProbeResult.success(
+                        KnownSecurityVulnerabilityProbe.KEY, warningId + "|" + url, probe.getVersion()));
     }
 
     @Test
@@ -169,6 +167,7 @@ class KnownSecurityVulnerabilityProbeTest extends AbstractProbeTest<KnownSecurit
         final String pluginName = "foo-bar";
         final VersionNumber pluginVersion = new VersionNumber("1.1");
         final String warningId = "SECURITY-1";
+        final String url = "http://link-to-issue";
 
         final var plugin = mock(io.jenkins.pluginhealth.scoring.model.Plugin.class);
         final ProbeContext ctx = mock(ProbeContext.class);
@@ -176,22 +175,20 @@ class KnownSecurityVulnerabilityProbeTest extends AbstractProbeTest<KnownSecurit
 
         when(plugin.getName()).thenReturn(pluginName);
         when(plugin.getVersion()).thenReturn(pluginVersion);
-        when(ctx.getUpdateCenter()).thenReturn(
-            new UpdateCenter(
-                Collections.emptyMap(),
-                Collections.emptyMap(),
-                List.of(
-                    new SecurityWarning(warningId, pluginName, List.of(new SecurityWarningVersion(null, ".*")))
-                )
-            )
-        );
+        when(ctx.getUpdateCenter())
+                .thenReturn(new UpdateCenter(
+                        Collections.emptyMap(),
+                        Collections.emptyMap(),
+                        List.of(new SecurityWarning(
+                                warningId, pluginName, url, List.of(new SecurityWarningVersion(null, ".*"))))));
 
         final ProbeResult result = probe.apply(plugin, ctx);
         assertThat(result)
-            .isNotNull()
-            .usingRecursiveComparison()
-            .comparingOnlyFields("id", "status", "message")
-            .isEqualTo(ProbeResult.success(KnownSecurityVulnerabilityProbe.KEY, warningId, probe.getVersion()));
+                .isNotNull()
+                .usingRecursiveComparison()
+                .comparingOnlyFields("id", "status", "message")
+                .isEqualTo(ProbeResult.success(
+                        KnownSecurityVulnerabilityProbe.KEY, warningId + "|" + url, probe.getVersion()));
     }
 
     @Test
@@ -200,6 +197,8 @@ class KnownSecurityVulnerabilityProbeTest extends AbstractProbeTest<KnownSecurit
         final VersionNumber pluginVersion = new VersionNumber("1.1");
         final String warningId1 = "SECURITY-1";
         final String warningId2 = "SECURITY-2";
+        final String url1 = "http://link-to-issue";
+        final String url2 = "https://link-to-other-issue";
 
         final var plugin = mock(io.jenkins.pluginhealth.scoring.model.Plugin.class);
         final ProbeContext ctx = mock(ProbeContext.class);
@@ -207,23 +206,28 @@ class KnownSecurityVulnerabilityProbeTest extends AbstractProbeTest<KnownSecurit
 
         when(plugin.getName()).thenReturn(pluginName);
         when(plugin.getVersion()).thenReturn(pluginVersion);
-        when(ctx.getUpdateCenter()).thenReturn(
-            new UpdateCenter(
-                Collections.emptyMap(),
-                Collections.emptyMap(),
-                List.of(
-                    new SecurityWarning(warningId1, pluginName, List.of(new SecurityWarningVersion(null, ".*"))),
-                    new SecurityWarning(warningId2, pluginName, List.of(new SecurityWarningVersion(null, ".*")))
-                )
-            )
-        );
+        when(ctx.getUpdateCenter())
+                .thenReturn(new UpdateCenter(
+                        Collections.emptyMap(),
+                        Collections.emptyMap(),
+                        List.of(
+                                new SecurityWarning(
+                                        warningId1, pluginName, url1, List.of(new SecurityWarningVersion(null, ".*"))),
+                                new SecurityWarning(
+                                        warningId2,
+                                        pluginName,
+                                        url2,
+                                        List.of(new SecurityWarningVersion(null, ".*"))))));
 
         final ProbeResult result = probe.apply(plugin, ctx);
         assertThat(result)
-            .isNotNull()
-            .usingRecursiveComparison()
-            .comparingOnlyFields("id", "status", "message")
-            .isEqualTo(ProbeResult.success(KnownSecurityVulnerabilityProbe.KEY, "%s, %s".formatted(warningId1, warningId2), probe.getVersion()));
+                .isNotNull()
+                .usingRecursiveComparison()
+                .comparingOnlyFields("id", "status", "message")
+                .isEqualTo(ProbeResult.success(
+                        KnownSecurityVulnerabilityProbe.KEY,
+                        "%s, %s".formatted(warningId1 + "|" + url1, warningId2 + "|" + url2),
+                        probe.getVersion()));
     }
 
     @Test
@@ -232,6 +236,8 @@ class KnownSecurityVulnerabilityProbeTest extends AbstractProbeTest<KnownSecurit
         final VersionNumber pluginVersion = new VersionNumber("1.2");
         final String warningId1 = "SECURITY-1";
         final String warningId2 = "SECURITY-1";
+        final String url1 = "http://link-to-issue";
+        final String url2 = "https://link-to-other-issue";
 
         final var plugin = mock(io.jenkins.pluginhealth.scoring.model.Plugin.class);
         final ProbeContext ctx = mock(ProbeContext.class);
@@ -239,23 +245,26 @@ class KnownSecurityVulnerabilityProbeTest extends AbstractProbeTest<KnownSecurit
 
         when(plugin.getName()).thenReturn(pluginName);
         when(plugin.getVersion()).thenReturn(pluginVersion);
-        when(ctx.getUpdateCenter()).thenReturn(
-            new UpdateCenter(
-                Collections.emptyMap(),
-                Collections.emptyMap(),
-                List.of(
-                    new SecurityWarning(warningId1, pluginName, List.of(new SecurityWarningVersion(null, ".*"))),
-                    new SecurityWarning(warningId2, pluginName, List.of(new SecurityWarningVersion(new VersionNumber("1.1"), ".*")))
-                )
-            )
-        );
+        when(ctx.getUpdateCenter())
+                .thenReturn(new UpdateCenter(
+                        Collections.emptyMap(),
+                        Collections.emptyMap(),
+                        List.of(
+                                new SecurityWarning(
+                                        warningId1, pluginName, url1, List.of(new SecurityWarningVersion(null, ".*"))),
+                                new SecurityWarning(
+                                        warningId2,
+                                        pluginName,
+                                        url2,
+                                        List.of(new SecurityWarningVersion(new VersionNumber("1.1"), ".*"))))));
 
         final ProbeResult result = probe.apply(plugin, ctx);
         assertThat(result)
-            .isNotNull()
-            .usingRecursiveComparison()
-            .comparingOnlyFields("id", "status", "message")
-            .isEqualTo(ProbeResult.success(KnownSecurityVulnerabilityProbe.KEY, warningId1, probe.getVersion()));
+                .isNotNull()
+                .usingRecursiveComparison()
+                .comparingOnlyFields("id", "status", "message")
+                .isEqualTo(ProbeResult.success(
+                        KnownSecurityVulnerabilityProbe.KEY, warningId1 + "|" + url1, probe.getVersion()));
     }
 
     @Test
@@ -269,21 +278,22 @@ class KnownSecurityVulnerabilityProbeTest extends AbstractProbeTest<KnownSecurit
 
         when(plugin.getName()).thenReturn(pluginName);
         when(plugin.getVersion()).thenReturn(pluginVersion);
-        when(ctx.getUpdateCenter()).thenReturn(
-            new UpdateCenter(
-                Collections.emptyMap(),
-                Collections.emptyMap(),
-                List.of(
-                    new SecurityWarning("SECURITY-1", pluginName, List.of(new SecurityWarningVersion(null, "[0-1]\\..*")))
-                )
-            )
-        );
+        when(ctx.getUpdateCenter())
+                .thenReturn(new UpdateCenter(
+                        Collections.emptyMap(),
+                        Collections.emptyMap(),
+                        List.of(new SecurityWarning(
+                                "SECURITY-1",
+                                pluginName,
+                                "http://link-to-issue",
+                                List.of(new SecurityWarningVersion(null, "[0-1]\\..*"))))));
 
         final ProbeResult result = probe.apply(plugin, ctx);
         assertThat(result)
-            .isNotNull()
-            .usingRecursiveComparison()
-            .comparingOnlyFields("id", "status", "message")
-            .isEqualTo(ProbeResult.success(KnownSecurityVulnerabilityProbe.KEY, "No known security vulnerabilities.", probe.getVersion()));
+                .isNotNull()
+                .usingRecursiveComparison()
+                .comparingOnlyFields("id", "status", "message")
+                .isEqualTo(ProbeResult.success(
+                        KnownSecurityVulnerabilityProbe.KEY, "No known security vulnerabilities.", probe.getVersion()));
     }
 }
