@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c) 2023 Jenkins Infra
+ * Copyright (c) 2023-2024 Jenkins Infra
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -21,7 +21,6 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-
 package io.jenkins.pluginhealth.scoring.probes;
 
 import java.util.List;
@@ -46,21 +45,19 @@ public class KnownSecurityVulnerabilityProbe extends Probe {
     protected ProbeResult doApply(Plugin plugin, ProbeContext context) {
         final List<SecurityWarning> warnings = context.getUpdateCenter().warnings();
         final String issues = warnings.stream()
-            .filter(w -> w.name().equals(plugin.getName()))
-            .filter(w -> w.versions().stream().anyMatch(securityWarningVersion -> {
-                if (securityWarningVersion.lastVersion() != null) {
-                    return plugin.getVersion().isOlderThanOrEqualTo(securityWarningVersion.lastVersion());
-                }
-                final Pattern pattern = Pattern.compile(securityWarningVersion.pattern());
-                final Matcher matcher = pattern.matcher(plugin.getVersion().toString());
-                return matcher.find();
-            }))
-            .map(SecurityWarning::id)
-            .collect(Collectors.joining(", "));
+                .filter(w -> w.name().equals(plugin.getName()))
+                .filter(w -> w.versions().stream().anyMatch(securityWarningVersion -> {
+                    if (securityWarningVersion.lastVersion() != null) {
+                        return plugin.getVersion().isOlderThanOrEqualTo(securityWarningVersion.lastVersion());
+                    }
+                    final Pattern pattern = Pattern.compile(securityWarningVersion.pattern());
+                    final Matcher matcher = pattern.matcher(plugin.getVersion().toString());
+                    return matcher.find();
+                }))
+                .map(warning -> String.join("|", warning.id(), warning.url()))
+                .collect(Collectors.joining(", "));
 
-        return !issues.isBlank() ?
-            this.success(issues) :
-            this.success("No known security vulnerabilities.");
+        return !issues.isBlank() ? this.success(issues) : this.success("No known security vulnerabilities.");
     }
 
     @Override
@@ -75,6 +72,6 @@ public class KnownSecurityVulnerabilityProbe extends Probe {
 
     @Override
     public long getVersion() {
-        return 1;
+        return 2;
     }
 }
