@@ -21,7 +21,6 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-
 package io.jenkins.pluginhealth.scoring.probes;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -54,13 +53,16 @@ public class PluginDescriptionMigrationProbeTest extends AbstractProbeTest<Plugi
         final Plugin plugin = mock(Plugin.class);
         final ProbeContext ctx = mock(ProbeContext.class);
 
+        when(plugin.getName()).thenReturn("plugin-a");
+
         final PluginDescriptionMigrationProbe probe = getSpy();
         final ProbeResult result = probe.apply(plugin, ctx);
 
         assertThat(result)
-            .usingRecursiveComparison()
-            .comparingOnlyFields("id", "status", "message")
-            .isEqualTo(ProbeResult.error(PluginDescriptionMigrationProbe.KEY, "Cannot access plugin repository.", 0));
+                .usingRecursiveComparison()
+                .comparingOnlyFields("id", "status", "message")
+                .isEqualTo(ProbeResult.error(
+                        PluginDescriptionMigrationProbe.KEY, "There is no local repository for plugin plugin-a.", 0));
     }
 
     @Test
@@ -75,9 +77,10 @@ public class PluginDescriptionMigrationProbeTest extends AbstractProbeTest<Plugi
         final ProbeResult result = probe.apply(plugin, ctx);
 
         assertThat(result)
-            .usingRecursiveComparison()
-            .comparingOnlyFields("id", "message", "status")
-            .isEqualTo(ProbeResult.error(PluginDescriptionMigrationProbe.KEY, "Cannot browse plugin source code folder.", 0));
+                .usingRecursiveComparison()
+                .comparingOnlyFields("id", "message", "status")
+                .isEqualTo(ProbeResult.error(
+                        PluginDescriptionMigrationProbe.KEY, "Cannot browse plugin source code folder.", 0));
     }
 
     @Test
@@ -93,9 +96,12 @@ public class PluginDescriptionMigrationProbeTest extends AbstractProbeTest<Plugi
         final ProbeResult result = probe.apply(plugin, ctx);
 
         assertThat(result)
-            .usingRecursiveComparison()
-            .comparingOnlyFields("id", "message", "status")
-            .isEqualTo(ProbeResult.success(PluginDescriptionMigrationProbe.KEY, "There is no `index.jelly` file in `src/main/resources`.", 0));
+                .usingRecursiveComparison()
+                .comparingOnlyFields("id", "message", "status")
+                .isEqualTo(ProbeResult.success(
+                        PluginDescriptionMigrationProbe.KEY,
+                        "There is no `index.jelly` file in `src/main/resources`.",
+                        0));
     }
 
     @Test
@@ -104,7 +110,8 @@ public class PluginDescriptionMigrationProbeTest extends AbstractProbeTest<Plugi
         final ProbeContext ctx = mock(ProbeContext.class);
 
         final Path repository = Files.createTempDirectory(getClass().getSimpleName());
-        final Path resources = Files.createDirectories(repository.resolve("src").resolve("main").resolve("resources"));
+        final Path resources = Files.createDirectories(
+                repository.resolve("src").resolve("main").resolve("resources"));
         final Path jelly = Files.createFile(resources.resolve("index.jelly"));
         Files.writeString(jelly, """
             <div>
@@ -117,9 +124,12 @@ public class PluginDescriptionMigrationProbeTest extends AbstractProbeTest<Plugi
         final ProbeResult result = probe.apply(plugin, ctx);
 
         assertThat(result)
-            .usingRecursiveComparison()
-            .comparingOnlyFields("id", "message", "status")
-            .isEqualTo(ProbeResult.success(PluginDescriptionMigrationProbe.KEY, "Plugin is using description from the plugin archetype.", 0));
+                .usingRecursiveComparison()
+                .comparingOnlyFields("id", "message", "status")
+                .isEqualTo(ProbeResult.success(
+                        PluginDescriptionMigrationProbe.KEY,
+                        "Plugin is using description from the plugin archetype.",
+                        0));
     }
 
     @Test
@@ -129,7 +139,8 @@ public class PluginDescriptionMigrationProbeTest extends AbstractProbeTest<Plugi
 
         final Path repository = Files.createTempDirectory(getClass().getSimpleName());
         final Path module = Files.createDirectory(repository.resolve("plugin"));
-        final Path resources = Files.createDirectories(module.resolve("src").resolve("main").resolve("resources"));
+        final Path resources =
+                Files.createDirectories(module.resolve("src").resolve("main").resolve("resources"));
         final Path jelly = Files.createFile(resources.resolve("index.jelly"));
         Files.writeString(jelly, """
             <div>
@@ -143,9 +154,12 @@ public class PluginDescriptionMigrationProbeTest extends AbstractProbeTest<Plugi
         final ProbeResult result = probe.apply(plugin, ctx);
 
         assertThat(result)
-            .usingRecursiveComparison()
-            .comparingOnlyFields("id", "message", "status")
-            .isEqualTo(ProbeResult.success(PluginDescriptionMigrationProbe.KEY, "Plugin is using description from the plugin archetype.", 0));
+                .usingRecursiveComparison()
+                .comparingOnlyFields("id", "message", "status")
+                .isEqualTo(ProbeResult.success(
+                        PluginDescriptionMigrationProbe.KEY,
+                        "Plugin is using description from the plugin archetype.",
+                        0));
     }
 
     @Test
@@ -154,9 +168,12 @@ public class PluginDescriptionMigrationProbeTest extends AbstractProbeTest<Plugi
         final ProbeContext ctx = mock(ProbeContext.class);
 
         final Path repository = Files.createTempDirectory(getClass().getSimpleName());
-        final Path resources = Files.createDirectories(repository.resolve("src").resolve("main").resolve("resources"));
+        final Path resources = Files.createDirectories(
+                repository.resolve("src").resolve("main").resolve("resources"));
         final Path jelly = Files.createFile(resources.resolve("index.jelly"));
-        Files.writeString(jelly, """
+        Files.writeString(
+                jelly,
+                """
             <div>
                 This is a plugin doing something.
             </div>
@@ -167,9 +184,10 @@ public class PluginDescriptionMigrationProbeTest extends AbstractProbeTest<Plugi
         final ProbeResult result = probe.apply(plugin, ctx);
 
         assertThat(result)
-            .usingRecursiveComparison()
-            .comparingOnlyFields("id", "message", "status")
-            .isEqualTo(ProbeResult.success(PluginDescriptionMigrationProbe.KEY, "Plugin seems to have a correct description.", 0));
+                .usingRecursiveComparison()
+                .comparingOnlyFields("id", "message", "status")
+                .isEqualTo(ProbeResult.success(
+                        PluginDescriptionMigrationProbe.KEY, "Plugin seems to have a correct description.", 0));
     }
 
     @Test
@@ -179,9 +197,12 @@ public class PluginDescriptionMigrationProbeTest extends AbstractProbeTest<Plugi
 
         final Path repository = Files.createTempDirectory(getClass().getSimpleName());
         final Path module = Files.createDirectory(repository.resolve("plugin"));
-        final Path resources = Files.createDirectories(module.resolve("src").resolve("main").resolve("resources"));
+        final Path resources =
+                Files.createDirectories(module.resolve("src").resolve("main").resolve("resources"));
         final Path jelly = Files.createFile(resources.resolve("index.jelly"));
-        Files.writeString(jelly, """
+        Files.writeString(
+                jelly,
+                """
             <div>
                 This is a plugin doing something.
             </div>
@@ -193,8 +214,9 @@ public class PluginDescriptionMigrationProbeTest extends AbstractProbeTest<Plugi
         final ProbeResult result = probe.apply(plugin, ctx);
 
         assertThat(result)
-            .usingRecursiveComparison()
-            .comparingOnlyFields("id", "message", "status")
-            .isEqualTo(ProbeResult.success(PluginDescriptionMigrationProbe.KEY, "Plugin seems to have a correct description.", 0));
+                .usingRecursiveComparison()
+                .comparingOnlyFields("id", "message", "status")
+                .isEqualTo(ProbeResult.success(
+                        PluginDescriptionMigrationProbe.KEY, "Plugin seems to have a correct description.", 0));
     }
 }
