@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c) 2023-2024 Jenkins Infra
+ * Copyright (c) 2024 Jenkins Infra
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -21,8 +21,28 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package io.jenkins.pluginhealth.scoring.model.updatecenter;
+package io.jenkins.pluginhealth.scoring.probes;
 
-import java.util.List;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
 
-public record SecurityWarning(String id, String name, String url, List<SecurityWarningVersion> versions) {}
+import java.util.Optional;
+
+import io.jenkins.pluginhealth.scoring.model.Plugin;
+import io.jenkins.pluginhealth.scoring.model.updatecenter.UpdateCenter;
+
+import org.junit.jupiter.api.Test;
+
+public class ProbeContextTest {
+    @Test
+    void shouldBeAbleToReturnCorrectPluginRepositoryName() throws Exception {
+        final Plugin plugin = mock(Plugin.class);
+        final UpdateCenter uc = mock(UpdateCenter.class);
+
+        when(plugin.getScm()).thenReturn("https://github.com/jenkinsci/git-client-plugin");
+
+        final ProbeContext ctx = new ProbeContext(plugin, uc);
+        assertThat(ctx.getRepositoryName()).isEqualTo(Optional.of("jenkinsci/git-client-plugin"));
+    }
+}
