@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c) 2023 Jenkins Infra
+ * Copyright (c) 2023-2024 Jenkins Infra
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -21,7 +21,6 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-
 package io.jenkins.pluginhealth.scoring.scores;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -29,10 +28,12 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.when;
 
+import java.util.List;
 import java.util.Map;
 
 import io.jenkins.pluginhealth.scoring.model.Plugin;
 import io.jenkins.pluginhealth.scoring.model.ProbeResult;
+import io.jenkins.pluginhealth.scoring.model.Resolution;
 import io.jenkins.pluginhealth.scoring.model.ScoreResult;
 import io.jenkins.pluginhealth.scoring.probes.KnownSecurityVulnerabilityProbe;
 
@@ -45,19 +46,59 @@ class SecurityWarningScoringTest extends AbstractScoringTest<SecurityWarningScor
     }
 
     @Test
-    void shouldBeAbleToDetectPluginWithSecurityWarning() {
+    void shouldBeAbleToDetectPluginWithSingleSecurityWarning() {
         final Plugin plugin = mock(Plugin.class);
         final SecurityWarningScoring scoring = getSpy();
 
-        when(plugin.getDetails()).thenReturn(Map.of(
-            KnownSecurityVulnerabilityProbe.KEY, ProbeResult.success(KnownSecurityVulnerabilityProbe.KEY, "SECURITY-123, link-to-security-advisory", 1)
-        ));
+        when(plugin.getDetails())
+                .thenReturn(Map.of(
+                        KnownSecurityVulnerabilityProbe.KEY,
+                        ProbeResult.success(
+                                KnownSecurityVulnerabilityProbe.KEY,
+                                "SECURITY-123|https://link-to-security-advisory",
+                                1)));
 
         final ScoreResult result = scoring.apply(plugin);
 
         assertThat(result.key()).isEqualTo("security");
         assertThat(result.weight()).isEqualTo(1f);
         assertThat(result.value()).isEqualTo(0);
+        result.componentsResults().forEach(ScoringComponentResult -> {
+            ScoringComponentResult.resolutions().forEach(resolution -> {
+                assertThat(resolution.link()).isEqualTo("https://link-to-security-advisory");
+                assertThat(resolution.text()).isEqualTo("SECURITY-123");
+            });
+        });
+    }
+
+    @Test
+    void shouldBeAbleToDetectPluginWithMultipleSecurityWarning() {
+        final Plugin plugin = mock(Plugin.class);
+        final SecurityWarningScoring scoring = getSpy();
+        final List<String> securityInfo = List.of(
+                "SECURITY-123|https://link-to-security-advisory", "SECURITY-123|https://link-to-security-advisory");
+        when(plugin.getDetails())
+                .thenReturn(Map.of(
+                        KnownSecurityVulnerabilityProbe.KEY,
+                        ProbeResult.success(KnownSecurityVulnerabilityProbe.KEY, String.join(", ", securityInfo), 1)));
+
+        final ScoreResult result = scoring.apply(plugin);
+
+        assertThat(result.key()).isEqualTo("security");
+        assertThat(result.weight()).isEqualTo(1f);
+        assertThat(result.value()).isEqualTo(0);
+        final List<Resolution> resolutionList = securityInfo.stream()
+                .map(securityMessage -> {
+                    String[] securityDetails = securityMessage.split("\\|");
+                    return new Resolution(securityDetails[0].trim(), securityDetails[1]);
+                })
+                .toList();
+
+        result.componentsResults().forEach(componentResult -> {
+            componentResult.resolutions().forEach(resolution -> {
+                assert resolutionList.contains(resolution);
+            });
+        });
     }
 
     @Test
@@ -79,9 +120,11 @@ class SecurityWarningScoringTest extends AbstractScoringTest<SecurityWarningScor
         final Plugin plugin = mock(Plugin.class);
         final SecurityWarningScoring scoring = getSpy();
 
-        when(plugin.getDetails()).thenReturn(Map.of(
-            KnownSecurityVulnerabilityProbe.KEY, ProbeResult.success(KnownSecurityVulnerabilityProbe.KEY, "No known security vulnerabilities.", 1)
-        ));
+        when(plugin.getDetails())
+                .thenReturn(Map.of(
+                        KnownSecurityVulnerabilityProbe.KEY,
+                        ProbeResult.success(
+                                KnownSecurityVulnerabilityProbe.KEY, "No known security vulnerabilities.", 2)));
 
         final ScoreResult result = scoring.apply(plugin);
 
