@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c) 2023 Jenkins Infra
+ * Copyright (c) 2023-2024 Jenkins Infra
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -21,7 +21,6 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-
 package io.jenkins.pluginhealth.scoring.probes;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -59,19 +58,24 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(MockitoExtension.class)
 class ProbeEngineTest {
-    @Mock private PluginService pluginService;
-    @Mock private ProbeService probeService;
-    @Mock private UpdateCenterService updateCenterService;
-    @Mock private GitHub gitHub;
-    @Mock private PluginDocumentationService pluginDocumentationService;
+    @Mock
+    private PluginService pluginService;
+
+    @Mock
+    private ProbeService probeService;
+
+    @Mock
+    private UpdateCenterService updateCenterService;
+
+    @Mock
+    private GitHub gitHub;
+
+    @Mock
+    private PluginDocumentationService pluginDocumentationService;
 
     @BeforeEach
     void setup() throws IOException {
-        when(updateCenterService.fetchUpdateCenter()).thenReturn(new UpdateCenter(
-            Map.of(),
-            Map.of(),
-            List.of()
-        ));
+        when(updateCenterService.fetchUpdateCenter()).thenReturn(new UpdateCenter(Map.of(), Map.of(), List.of()));
     }
 
     @Test
@@ -82,16 +86,17 @@ class ProbeEngineTest {
 
         final ProbeResult expectedResult = ProbeResult.success("foo", "bar", 1);
 
-        when(plugin.getScm()).thenReturn("this-is-ok-for-testing");
         when(plugin.getDetails()).thenReturn(Map.of());
         when(probe.key()).thenReturn("probe");
         when(probe.doApply(plugin, ctx)).thenReturn(expectedResult);
 
-        when(probeService.getProbeContext(any(Plugin.class), any(UpdateCenter.class))).thenReturn(ctx);
+        when(probeService.getProbeContext(any(Plugin.class), any(UpdateCenter.class)))
+                .thenReturn(ctx);
         when(probeService.getProbes()).thenReturn(List.of(probe));
         when(pluginService.streamAll()).thenReturn(Stream.of(plugin));
 
-        final ProbeEngine probeEngine = new ProbeEngine(probeService, pluginService, updateCenterService, gitHub, pluginDocumentationService);
+        final ProbeEngine probeEngine =
+                new ProbeEngine(probeService, pluginService, updateCenterService, gitHub, pluginDocumentationService);
         probeEngine.run();
 
         verify(probe).doApply(plugin, ctx);
@@ -106,7 +111,6 @@ class ProbeEngineTest {
         final Probe probe = spy(Probe.class);
         final ProbeContext ctx = mock(ProbeContext.class);
 
-        when(plugin.getScm()).thenReturn("this-is-ok-for-testing");
         when(plugin.getName()).thenReturn("foo");
         when(plugin.getReleaseTimestamp()).thenReturn(ZonedDateTime.now().minusDays(1));
         when(plugin.getDetails()).thenReturn(Map.of(probeKey, ProbeResult.success(probeKey, "This is good", 1)));
@@ -115,11 +119,13 @@ class ProbeEngineTest {
         when(probe.key()).thenReturn(probeKey);
         when(probe.getVersion()).thenReturn(1L);
 
-        when(probeService.getProbeContext(any(Plugin.class), any(UpdateCenter.class))).thenReturn(ctx);
+        when(probeService.getProbeContext(any(Plugin.class), any(UpdateCenter.class)))
+                .thenReturn(ctx);
         when(probeService.getProbes()).thenReturn(List.of(probe));
         when(pluginService.streamAll()).thenReturn(Stream.of(plugin));
 
-        final ProbeEngine probeEngine = new ProbeEngine(probeService, pluginService, updateCenterService, gitHub, pluginDocumentationService);
+        final ProbeEngine probeEngine =
+                new ProbeEngine(probeService, pluginService, updateCenterService, gitHub, pluginDocumentationService);
         probeEngine.run();
 
         verify(probe, never()).doApply(plugin, ctx);
@@ -132,10 +138,15 @@ class ProbeEngineTest {
         final Probe probe = spy(Probe.class);
         final ProbeContext ctx = mock(ProbeContext.class);
 
-        when(plugin.getScm()).thenReturn("this-is-ok-for-testing");
-        when(plugin.getDetails()).thenReturn(Map.of(
-            "probe", new ProbeResult("probe", "message", ProbeResult.Status.SUCCESS, ZonedDateTime.now().minusDays(1), 1)
-        ));
+        when(plugin.getDetails())
+                .thenReturn(Map.of(
+                        "probe",
+                        new ProbeResult(
+                                "probe",
+                                "message",
+                                ProbeResult.Status.SUCCESS,
+                                ZonedDateTime.now().minusDays(1),
+                                1)));
         when(plugin.getName()).thenReturn("foo");
 
         when(probe.getVersion()).thenReturn(1L);
@@ -143,11 +154,13 @@ class ProbeEngineTest {
         when(probe.requiresRelease()).thenReturn(false);
         when(probe.isSourceCodeRelated()).thenReturn(true);
 
-        when(probeService.getProbeContext(any(Plugin.class), any(UpdateCenter.class))).thenReturn(ctx);
+        when(probeService.getProbeContext(any(Plugin.class), any(UpdateCenter.class)))
+                .thenReturn(ctx);
         when(probeService.getProbes()).thenReturn(List.of(probe));
         when(pluginService.streamAll()).thenReturn(Stream.of(plugin));
 
-        final ProbeEngine probeEngine = new ProbeEngine(probeService, pluginService, updateCenterService, gitHub, pluginDocumentationService);
+        final ProbeEngine probeEngine =
+                new ProbeEngine(probeService, pluginService, updateCenterService, gitHub, pluginDocumentationService);
         probeEngine.run();
 
         verify(probe, never()).doApply(plugin, ctx);
@@ -162,24 +175,33 @@ class ProbeEngineTest {
 
         final ProbeResult result = new ProbeResult(probeKey, "message", ProbeResult.Status.SUCCESS, 1);
 
-        when(plugin.getScm()).thenReturn("this-is-ok-for-testing");
-        when(plugin.getDetails()).thenReturn(Map.of(
-            probeKey, new ProbeResult(probeKey, "message", ProbeResult.Status.SUCCESS, ZonedDateTime.now().minusDays(1), 1)
-        ));
+        when(plugin.getDetails())
+                .thenReturn(Map.of(
+                        probeKey,
+                        new ProbeResult(
+                                probeKey,
+                                "message",
+                                ProbeResult.Status.SUCCESS,
+                                ZonedDateTime.now().minusDays(1),
+                                1)));
 
         when(probe.getVersion()).thenReturn(1L);
-        when(ctx.getScmRepository()).thenReturn(Optional.of(Files.createTempDirectory("ProbeEngineTest-shouldApplyProbeRelatedToCodeWithNewCommit")));
+        when(ctx.getScmRepository())
+                .thenReturn(Optional.of(
+                        Files.createTempDirectory("ProbeEngineTest-shouldApplyProbeRelatedToCodeWithNewCommit")));
         when(ctx.getLastCommitDate()).thenReturn(Optional.of(ZonedDateTime.now()));
 
         when(probe.key()).thenReturn(probeKey);
         when(probe.isSourceCodeRelated()).thenReturn(true);
         when(probe.doApply(plugin, ctx)).thenReturn(result);
 
-        when(probeService.getProbeContext(any(Plugin.class), any(UpdateCenter.class))).thenReturn(ctx);
+        when(probeService.getProbeContext(any(Plugin.class), any(UpdateCenter.class)))
+                .thenReturn(ctx);
         when(probeService.getProbes()).thenReturn(List.of(probe));
         when(pluginService.streamAll()).thenReturn(Stream.of(plugin));
 
-        final ProbeEngine probeEngine = new ProbeEngine(probeService, pluginService, updateCenterService, gitHub, pluginDocumentationService);
+        final ProbeEngine probeEngine =
+                new ProbeEngine(probeService, pluginService, updateCenterService, gitHub, pluginDocumentationService);
         probeEngine.run();
 
         verify(probe).doApply(plugin, ctx);
@@ -193,20 +215,29 @@ class ProbeEngineTest {
         final Probe probe = spy(Probe.class);
         final ProbeContext ctx = mock(ProbeContext.class);
 
-        when(plugin.getScm()).thenReturn("this-is-ok-for-testing");
         when(plugin.getReleaseTimestamp()).thenReturn(ZonedDateTime.now());
-        when(plugin.getDetails()).thenReturn(Map.of(probeKey, new ProbeResult(probeKey, "this is ok", ProbeResult.Status.SUCCESS, ZonedDateTime.now().minusDays(1), 1)));
+        when(plugin.getDetails())
+                .thenReturn(Map.of(
+                        probeKey,
+                        new ProbeResult(
+                                probeKey,
+                                "this is ok",
+                                ProbeResult.Status.SUCCESS,
+                                ZonedDateTime.now().minusDays(1),
+                                1)));
 
         when(probe.getVersion()).thenReturn(1L);
         when(probe.key()).thenReturn(probeKey);
         when(probe.requiresRelease()).thenReturn(true);
         when(probe.doApply(plugin, ctx)).thenReturn(ProbeResult.success(probeKey, "This is also ok", 1));
 
-        when(probeService.getProbeContext(any(Plugin.class), any(UpdateCenter.class))).thenReturn(ctx);
+        when(probeService.getProbeContext(any(Plugin.class), any(UpdateCenter.class)))
+                .thenReturn(ctx);
         when(probeService.getProbes()).thenReturn(List.of(probe));
         when(pluginService.streamAll()).thenReturn(Stream.of(plugin));
 
-        final ProbeEngine probeEngine = new ProbeEngine(probeService, pluginService, updateCenterService, gitHub, pluginDocumentationService);
+        final ProbeEngine probeEngine =
+                new ProbeEngine(probeService, pluginService, updateCenterService, gitHub, pluginDocumentationService);
         probeEngine.run();
 
         verify(probe).doApply(eq(plugin), any(ProbeContext.class));
@@ -220,22 +251,28 @@ class ProbeEngineTest {
         final Probe probe = spy(Probe.class);
         final ProbeContext ctx = mock(ProbeContext.class);
 
-        when(plugin.getScm()).thenReturn("this-is-ok-for-testing");
-        when(plugin.getDetails()).thenReturn(Map.of(
-            probeKey,
-            new ProbeResult(probeKey, "this is ok", ProbeResult.Status.SUCCESS, ZonedDateTime.now().minusDays(1), 1)
-        ));
+        when(plugin.getDetails())
+                .thenReturn(Map.of(
+                        probeKey,
+                        new ProbeResult(
+                                probeKey,
+                                "this is ok",
+                                ProbeResult.Status.SUCCESS,
+                                ZonedDateTime.now().minusDays(1),
+                                1)));
 
         when(probe.getVersion()).thenReturn(1L);
         when(probe.requiresRelease()).thenReturn(false);
         when(probe.doApply(plugin, ctx)).thenReturn(ProbeResult.success(probeKey, "This is also ok", 1));
         when(probe.key()).thenReturn(probeKey);
 
-        when(probeService.getProbeContext(any(Plugin.class), any(UpdateCenter.class))).thenReturn(ctx);
+        when(probeService.getProbeContext(any(Plugin.class), any(UpdateCenter.class)))
+                .thenReturn(ctx);
         when(probeService.getProbes()).thenReturn(List.of(probe));
         when(pluginService.streamAll()).thenReturn(Stream.of(plugin));
 
-        final ProbeEngine probeEngine = new ProbeEngine(probeService, pluginService, updateCenterService, gitHub, pluginDocumentationService);
+        final ProbeEngine probeEngine =
+                new ProbeEngine(probeService, pluginService, updateCenterService, gitHub, pluginDocumentationService);
         probeEngine.run();
 
         verify(probe).doApply(plugin, ctx);
@@ -248,14 +285,15 @@ class ProbeEngineTest {
         final Probe probe = spy(Probe.class);
         final ProbeContext ctx = mock(ProbeContext.class);
 
-        when(plugin.getScm()).thenReturn("this-is-ok-for-testing");
         when(probe.doApply(plugin, ctx)).thenReturn(ProbeResult.error("foo", "bar", 1));
 
-        when(probeService.getProbeContext(any(Plugin.class), any(UpdateCenter.class))).thenReturn(ctx);
+        when(probeService.getProbeContext(any(Plugin.class), any(UpdateCenter.class)))
+                .thenReturn(ctx);
         when(probeService.getProbes()).thenReturn(List.of(probe));
         when(pluginService.streamAll()).thenReturn(Stream.of(plugin));
 
-        final ProbeEngine probeEngine = new ProbeEngine(probeService, pluginService, updateCenterService, gitHub, pluginDocumentationService);
+        final ProbeEngine probeEngine =
+                new ProbeEngine(probeService, pluginService, updateCenterService, gitHub, pluginDocumentationService);
         probeEngine.run();
 
         verify(plugin).addDetails(any(ProbeResult.class));
@@ -269,9 +307,9 @@ class ProbeEngineTest {
         final Probe probeTwo = new Probe() {
             @Override
             protected ProbeResult doApply(Plugin plugin, ProbeContext ctx) {
-                return plugin.getDetails().get("foo") != null ?
-                    ProbeResult.success(key(), "This is also ok", this.getVersion()) :
-                    ProbeResult.error(key(), "This cannot be validated", this.getVersion());
+                return plugin.getDetails().get("foo") != null
+                        ? ProbeResult.success(key(), "This is also ok", this.getVersion())
+                        : ProbeResult.error(key(), "This cannot be validated", this.getVersion());
             }
 
             @Override
@@ -293,13 +331,13 @@ class ProbeEngineTest {
         when(probeOne.key()).thenReturn("foo");
         when(probeOne.doApply(plugin, ctx)).thenReturn(ProbeResult.success("foo", "This is ok", 1));
 
-        when(probeService.getProbeContext(any(Plugin.class), any(UpdateCenter.class))).thenReturn(ctx);
+        when(probeService.getProbeContext(any(Plugin.class), any(UpdateCenter.class)))
+                .thenReturn(ctx);
         when(probeService.getProbes()).thenReturn(List.of(probeOne, probeTwo));
         when(pluginService.streamAll()).thenReturn(Stream.of(plugin));
 
-        when(plugin.getScm()).thenReturn("this-is-ok-for-testing");
-
-        final ProbeEngine probeEngine = new ProbeEngine(probeService, pluginService, updateCenterService, gitHub, pluginDocumentationService);
+        final ProbeEngine probeEngine =
+                new ProbeEngine(probeService, pluginService, updateCenterService, gitHub, pluginDocumentationService);
         probeEngine.run();
 
         verify(plugin, times(2)).addDetails(any(ProbeResult.class));
@@ -310,9 +348,8 @@ class ProbeEngineTest {
     void shouldForceProbeExecutionWhenNewVersionOfTheProbe() throws IOException {
         final String probeKey = "foo";
         final long version = 1L;
-        final ProbeResult newProbeResult = new ProbeResult(
-            probeKey, "this is a message", ProbeResult.Status.SUCCESS, version + 1
-        );
+        final ProbeResult newProbeResult =
+                new ProbeResult(probeKey, "this is a message", ProbeResult.Status.SUCCESS, version + 1);
 
         final Plugin plugin = mock(Plugin.class);
         final Probe probe = spy(new Probe() {
@@ -343,17 +380,17 @@ class ProbeEngineTest {
         });
         final ProbeContext ctx = mock(ProbeContext.class);
 
-        final ProbeResult previousResult = new ProbeResult(probeKey, "this is a message", ProbeResult.Status.SUCCESS, version);
-        when(plugin.getScm()).thenReturn("this-is-ok-for-testing");
-        when(plugin.getDetails()).thenReturn(Map.of(
-            probeKey, previousResult
-        ));
+        final ProbeResult previousResult =
+                new ProbeResult(probeKey, "this is a message", ProbeResult.Status.SUCCESS, version);
+        when(plugin.getDetails()).thenReturn(Map.of(probeKey, previousResult));
 
-        when(probeService.getProbeContext(any(Plugin.class), any(UpdateCenter.class))).thenReturn(ctx);
+        when(probeService.getProbeContext(any(Plugin.class), any(UpdateCenter.class)))
+                .thenReturn(ctx);
         when(probeService.getProbes()).thenReturn(List.of(probe));
         when(pluginService.streamAll()).thenReturn(Stream.of(plugin));
 
-        final ProbeEngine probeEngine = new ProbeEngine(probeService, pluginService, updateCenterService, gitHub, pluginDocumentationService);
+        final ProbeEngine probeEngine =
+                new ProbeEngine(probeService, pluginService, updateCenterService, gitHub, pluginDocumentationService);
         probeEngine.run();
 
         verify(probe).doApply(plugin, ctx);
@@ -367,20 +404,40 @@ class ProbeEngineTest {
         final Probe probe = mock(Probe.class);
         final ProbeContext ctx = mock(ProbeContext.class);
 
-        when(p1.getScm()).thenReturn("this-is-ok");
-        when(p2.getScm()).thenReturn("this-is-ok");
-
         when(probe.apply(p1, ctx)).thenReturn(ProbeResult.success("foo", "this is fine", 1));
         when(probe.apply(p2, ctx)).thenReturn(ProbeResult.success("foo", "this is ok too", 1));
 
         when(probeService.getProbes()).thenReturn(List.of(probe));
-        when(probeService.getProbeContext(any(Plugin.class), any(UpdateCenter.class))).thenReturn(ctx);
+        when(probeService.getProbeContext(any(Plugin.class), any(UpdateCenter.class)))
+                .thenReturn(ctx);
         when(pluginService.streamAll()).thenReturn(Stream.of(p1, p2));
 
-        final ProbeEngine probeEngine = new ProbeEngine(probeService, pluginService, updateCenterService, gitHub, pluginDocumentationService);
+        final ProbeEngine probeEngine =
+                new ProbeEngine(probeService, pluginService, updateCenterService, gitHub, pluginDocumentationService);
         probeEngine.run();
 
         verify(pluginDocumentationService).fetchPluginDocumentationUrl();
         verify(updateCenterService).fetchUpdateCenter();
+    }
+
+    @Test
+    void shouldNotPreventProbeExecutionWithNoSCM() throws Exception {
+        final Plugin plugin = mock(Plugin.class);
+        final Probe probe = spy(Probe.class);
+        final ProbeContext ctx = mock(ProbeContext.class);
+
+        when(probe.key()).thenReturn("probe");
+        when(probe.doApply(plugin, ctx)).thenReturn(ProbeResult.success("probe", "this is fine", 1));
+
+        when(probeService.getProbes()).thenReturn(List.of(probe));
+        when(probeService.getProbeContext(any(Plugin.class), any(UpdateCenter.class)))
+                .thenReturn(ctx);
+        when(pluginService.streamAll()).thenReturn(Stream.of(plugin));
+
+        final ProbeEngine probeEngine =
+                new ProbeEngine(probeService, pluginService, updateCenterService, gitHub, pluginDocumentationService);
+        probeEngine.run();
+
+        verify(probe).doApply(plugin, ctx);
     }
 }
