@@ -21,13 +21,14 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-
 package io.jenkins.pluginhealth.scoring.scores;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyMap;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
@@ -44,6 +45,7 @@ import io.jenkins.pluginhealth.scoring.model.Plugin;
 import io.jenkins.pluginhealth.scoring.model.ProbeResult;
 import io.jenkins.pluginhealth.scoring.model.Score;
 import io.jenkins.pluginhealth.scoring.model.ScoreResult;
+import io.jenkins.pluginhealth.scoring.model.ScoringComponentResult;
 import io.jenkins.pluginhealth.scoring.service.PluginService;
 import io.jenkins.pluginhealth.scoring.service.ScoreService;
 import io.jenkins.pluginhealth.scoring.service.ScoringService;
@@ -57,9 +59,14 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(MockitoExtension.class)
 class ScoringEngineTest {
-    @Mock private PluginService pluginService;
-    @Mock private ScoringService scoringService;
-    @Mock private ScoreService scoreService;
+    @Mock
+    private PluginService pluginService;
+
+    @Mock
+    private ScoringService scoringService;
+
+    @Mock
+    private ScoreService scoreService;
 
     @Test
     void shouldBeAbleToScoreOnePlugin() {
@@ -119,8 +126,9 @@ class ScoringEngineTest {
         final ArgumentCaptor<Score> scoreArgument = ArgumentCaptor.forClass(Score.class);
         verify(scoreService, times(3)).save(scoreArgument.capture());
         assertThat(scoreArgument.getAllValues())
-            .filteredOn(score -> Objects.nonNull(score) && score.getDetails().size() == 2 && score.getValue() == 85)
-            .hasSize(3);
+                .filteredOn(
+                        score -> Objects.nonNull(score) && score.getDetails().size() == 2 && score.getValue() == 85)
+                .hasSize(3);
     }
 
     @Test
@@ -128,9 +136,15 @@ class ScoringEngineTest {
         final Plugin pluginA = mock(Plugin.class);
         final String pluginName = "plugin-a";
         when(pluginA.getName()).thenReturn(pluginName);
-        when(pluginA.getDetails()).thenReturn(Map.of(
-            "foo-bar", new ProbeResult("foo-bar", "", ProbeResult.Status.SUCCESS, ZonedDateTime.now().minusMinutes(15), 1)
-        ));
+        when(pluginA.getDetails())
+                .thenReturn(Map.of(
+                        "foo-bar",
+                        new ProbeResult(
+                                "foo-bar",
+                                "",
+                                ProbeResult.Status.SUCCESS,
+                                ZonedDateTime.now().minusMinutes(15),
+                                1)));
 
         final String scoringAKey = "scoring-A";
         final Scoring scoringA = mock(Scoring.class);
@@ -139,9 +153,7 @@ class ScoringEngineTest {
 
         final Score oldPluginAScore = mock(Score.class);
         when(oldPluginAScore.getComputedAt()).thenReturn(ZonedDateTime.now().minusMinutes(5));
-        when(oldPluginAScore.getDetails()).thenReturn(Set.of(
-            new ScoreResult(scoringAKey, 100, 1, Set.of(), 1)
-        ));
+        when(oldPluginAScore.getDetails()).thenReturn(Set.of(new ScoreResult(scoringAKey, 100, 1, Set.of(), 1)));
 
         when(scoringService.getScoringList()).thenReturn(List.of(scoringA));
         when(scoreService.latestScoreFor(pluginName)).thenReturn(Optional.of(oldPluginAScore));
@@ -162,26 +174,26 @@ class ScoringEngineTest {
         final Score previousScore = mock(Score.class);
 
         when(plugin.getName()).thenReturn("foo-bar");
-        when(plugin.getDetails()).thenReturn(Map.of(
-            "foo-bar", new ProbeResult("foo-bar", "", ProbeResult.Status.SUCCESS, ZonedDateTime.now().minusDays(1), 1)
-        ));
+        when(plugin.getDetails())
+                .thenReturn(Map.of(
+                        "foo-bar",
+                        new ProbeResult(
+                                "foo-bar",
+                                "",
+                                ProbeResult.Status.SUCCESS,
+                                ZonedDateTime.now().minusDays(1),
+                                1)));
 
         when(scoringA.key()).thenReturn("scoring-a");
         when(scoringA.version()).thenReturn(2);
         final ScoreResult expectedNewScoreResult = new ScoreResult("scoring-a", 100, .5f, Set.of(), 2);
-        when(scoringA.apply(plugin)).thenReturn(
-            expectedNewScoreResult
-        );
+        when(scoringA.apply(plugin)).thenReturn(expectedNewScoreResult);
 
-        when(previousScore.getDetails()).thenReturn(Set.of(
-            new ScoreResult("scoring-a", 1, 1, Set.of(), 1)
-        ));
+        when(previousScore.getDetails()).thenReturn(Set.of(new ScoreResult("scoring-a", 1, 1, Set.of(), 1)));
         when(previousScore.getComputedAt()).thenReturn(ZonedDateTime.now().minusHours(1));
 
         when(scoringService.getScoringList()).thenReturn(List.of(scoringA));
-        when(scoreService.latestScoreFor("foo-bar")).thenReturn(Optional.of(
-            previousScore
-        ));
+        when(scoreService.latestScoreFor("foo-bar")).thenReturn(Optional.of(previousScore));
         when(scoreService.save(any(Score.class))).then(AdditionalAnswers.returnsFirstArg());
 
         final ScoringEngine scoringEngine = new ScoringEngine(scoringService, pluginService, scoreService);
@@ -205,18 +217,17 @@ class ScoringEngineTest {
 
         String pluginName = "plugin";
         when(plugin.getName()).thenReturn(pluginName);
-        when(plugin.getDetails()).thenReturn(Map.of(
-            "probe-a", new ProbeResult("probe-a", "", ProbeResult.Status.SUCCESS, now.minusMinutes(10), 1)
-        ));
+        when(plugin.getDetails())
+                .thenReturn(Map.of(
+                        "probe-a",
+                        new ProbeResult("probe-a", "", ProbeResult.Status.SUCCESS, now.minusMinutes(10), 1)));
 
         when(scoringA.version()).thenReturn(1);
         when(scoringA.key()).thenReturn("scoring-a");
         when(scoringB.key()).thenReturn("scoring-b");
 
         Score previousScore = mock(Score.class);
-        when(previousScore.getDetails()).thenReturn(Set.of(
-            new ScoreResult("scoring-a", 1, 1, Set.of(), 1)
-        ));
+        when(previousScore.getDetails()).thenReturn(Set.of(new ScoreResult("scoring-a", 1, 1, Set.of(), 1)));
         when(previousScore.getComputedAt()).thenReturn(now.minusMinutes(5));
         when(scoreService.latestScoreFor(pluginName)).thenReturn(Optional.of(previousScore));
 
@@ -239,9 +250,10 @@ class ScoringEngineTest {
 
         String pluginName = "plugin";
         when(plugin.getName()).thenReturn(pluginName);
-        when(plugin.getDetails()).thenReturn(Map.of(
-            "probe-a", new ProbeResult("probe-a", "", ProbeResult.Status.SUCCESS, now.minusMinutes(10), 1)
-        ));
+        when(plugin.getDetails())
+                .thenReturn(Map.of(
+                        "probe-a",
+                        new ProbeResult("probe-a", "", ProbeResult.Status.SUCCESS, now.minusMinutes(10), 1)));
 
         when(scoringA.version()).thenReturn(1);
         when(scoringA.key()).thenReturn("scoring-a");
@@ -249,10 +261,10 @@ class ScoringEngineTest {
         when(scoringB.key()).thenReturn("scoring-b");
 
         Score previousScore = mock(Score.class);
-        when(previousScore.getDetails()).thenReturn(Set.of(
-            new ScoreResult("scoring-a", 1, 1, Set.of(), 1),
-            new ScoreResult("scoring-b", 1, 1, Set.of(), 1)
-        ));
+        when(previousScore.getDetails())
+                .thenReturn(Set.of(
+                        new ScoreResult("scoring-a", 1, 1, Set.of(), 1),
+                        new ScoreResult("scoring-b", 1, 1, Set.of(), 1)));
         when(previousScore.getComputedAt()).thenReturn(now.minusMinutes(5));
         when(scoreService.latestScoreFor(pluginName)).thenReturn(Optional.of(previousScore));
 
@@ -263,5 +275,37 @@ class ScoringEngineTest {
 
         verify(scoringA).apply(plugin);
         verify(scoringB).apply(plugin);
+    }
+
+    @Test
+    void shouldNotPreventScoringWhenExceptionOnOnePlugin() {
+        final Plugin p1 = mock(Plugin.class);
+        final Plugin p2 = mock(Plugin.class);
+        final Scoring s1 = spy(Scoring.class);
+        final Scoring s2 = spy(Scoring.class);
+        final ScoringComponent sc1 = spy(ScoringComponent.class);
+        final ScoringComponent sc2 = spy(ScoringComponent.class);
+
+        when(p2.getName()).thenReturn("plugin-2");
+        when(s1.key()).thenReturn("scoring-a");
+        when(s2.key()).thenReturn("scoring-b");
+
+        when(s1.getComponents()).thenReturn(List.of(sc1));
+        when(s2.getComponents()).thenReturn(List.of(sc2));
+
+        when(sc1.getScore(p1, Map.of())).thenReturn(new ScoringComponentResult(100, 1, List.of()));
+        when(sc2.getScore(any(Plugin.class), anyMap())).thenReturn(new ScoringComponentResult(100, 1, List.of()));
+        when(sc1.getScore(p2, Map.of())).thenThrow(ArrayIndexOutOfBoundsException.class);
+
+        when(scoringService.getScoringList()).thenReturn(List.of(s1, s2));
+        when(pluginService.streamAll()).thenReturn(Stream.of(p1, p2));
+
+        final ScoringEngine scoringEngine = new ScoringEngine(scoringService, pluginService, scoreService);
+        scoringEngine.run();
+
+        verify(s1).apply(p1);
+        verify(s1).apply(p2);
+        verify(s2).apply(p1);
+        verify(s2).apply(p2);
     }
 }
